+++ CHANGELOG.md
@@ -2,7 +2,8 @@ v2.6.5 (FUTURE)
 
 ## Bug Fixes
 
-* [#3521](https://github.com/netbox-community/netbox/issues/3521) -  Fixed error in parseURL related to {{variables}} in API url
+* [#3519](https://github.com/netbox-community/netbox/issues/3519) -  Prevent cables from being terminated to virtual/wireless interfaces via API
+* [#3521](https://github.com/netbox-community/netbox/issues/3521) -  Fix error in `parseURL` related to variables in API URL
 
 v2.6.4 (2019-09-19)